backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
aiosqlite==0.20.0
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
httpx==0.27.2
cryptography==43.0.1
python-jose[cryptography]==3.3.0
celery==5.4.0
redis==5.0.8
langgraph==0.2.60
anthropic==0.40.0
langchain-anthropic==0.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
